working for small pi screen well, needs labels uncommented for larger screen size
